+++ shaders/import/at_tangent.glsl
@@ -1,0 +1,4 @@
+#define import_at_tangent
+#ifdef FIRST_PASS
+	vec4 at_tangent = vec4(0.0);
+#endif