+++ shaders/lib/lighting/fsh_lighting.glsl
@@ -300,7 +300,7 @@ void doFshLighting(inout vec3 color, float blockBrightness, float ambientBrightn
 		#include "/import/sunAngle.glsl"
 		#include "/import/ambientMoonPercent.glsl"
 		vec3 specularColor = sunAngle < 0.5 ? vec3(1.0, 1.0, 0.5) : vec3(0.5, 0.7, 0.9);
-		lighting += specularColor * specular * (0.2 + 0.75 * specular_amount) * shadowBrightness * (1.0 - 0.8 * ambientMoonPercent);
+		lighting += specularColor * specular * (0.2 + 0.7 * specular_amount) * shadowBrightness * (1.0 - 0.8 * ambientMoonPercent);
 	#endif
 	
 	float lightingBrightness = min(getLum(lighting), 1.0);