+++ shaders/program/gbuffers_water.glsl
@@ -1,3 +1,8 @@
+#if SHADOWS_ON_TRANSPARENTS == 0
+	#undef SHADOWS_ENABLED
+	#define SHADOWS_ENABLED 0
+#endif
+
 #ifdef FIRST_PASS
 	
 	in_out vec2 texcoord;
@@ -10,6 +15,9 @@
 	
 	flat in_out vec3 shadowcasterColor;
 	
+	#if WAVING_WATER_SURFACE_ENABLED == 1
+		in_out mat3 tbn;
+	#endif
 	#if BORDER_FOG_ENABLED == 1
 		in_out float fogAmount;
 	#endif
@@ -69,17 +77,22 @@ void main() {
 		
 		// waving water normals
 		#if WAVING_WATER_SURFACE_ENABLED == 1
-			#include "/import/frameTimeCounter.glsl"
+			float fresnel = -dot(normal, normalize(viewPos));
 			#include "/import/cameraPosition.glsl"
-			vec3 randomPoint = simplexNoise3From4(vec4((playerPos + cameraPosition) / WAVING_WATER_SCALE, frameTimeCounter * WAVING_WATER_SPEED));
-			randomPoint = normalize(randomPoint);
-			randomPoint += simplexNoise3From4(vec4((playerPos + cameraPosition) / WAVING_WATER_SCALE / 0.2, frameTimeCounter * WAVING_WATER_SPEED * 2.0)) * 0.5;
-			float wavingSurfaceAmount = mix(WAVING_WATER_SURFACE_AMOUNT_UNDERGROUND, WAVING_WATER_SURFACE_AMOUNT_SURFACE, lmcoord.y);
-			randomPoint = mix(normal, randomPoint, wavingSurfaceAmount);
-			randomPoint = normalize(randomPoint);
-			normal = normalize(normal + randomPoint * 0.05 * WAVING_WATER_NORMAL_AMOUNT * dot(normal, normalize(viewPos)));
-			float fresnel = dot(-randomPoint, normalize(viewPos));
-			color.rgb *= 1.0 - fresnel * fresnel * WAVING_WATER_FRESNEL_MULT * 0.5;
+			#include "/import/frameTimeCounter.glsl"
+			vec3 noisePos = vec3(playerPos.xz + cameraPosition.xz, frameTimeCounter * WAVING_WATER_SPEED);
+			noisePos.xy /= WAVING_WATER_SCALE * 2.0;
+			float wavingSurfaceAmount = mix(WAVING_WATER_SURFACE_AMOUNT_UNDERGROUND, WAVING_WATER_SURFACE_AMOUNT_SURFACE, lmcoord.y) * fresnel * 0.02;
+			float height = 1.0 - abs(simplexNoise(noisePos));
+			float heightX = 1.0 - abs(simplexNoise(noisePos + vec3(0.01, 0.0, 0.0)));
+			float heightZ = 1.0 - abs(simplexNoise(noisePos + vec3(0.0, 0.01, 0.0)));
+			vec3 dirX = vec3(0.01, (height - heightX) * wavingSurfaceAmount, 0.0);
+			vec3 dirZ = vec3(0.0, (height - heightZ) * wavingSurfaceAmount, 0.01);
+			normal = cross(dirZ, dirX);
+			normal = normalize(normal);
+			normal = tbn * vec3(-normal.x, normal.z, normal.y); // y = up -> z = up & tangent -> world
+			float newFresnel = dot(normal, normalize(viewPos)); // should be inverted but it would be inverted again in the next step anyways
+			color.rgb *= clamp(1.0 + WAVING_WATER_FRESNEL_MULT / wavingSurfaceAmount * 0.07 * (fresnel + newFresnel), 0.0, 1.5); // basically `color *= 1+(fresnel-newFresnel)` but it's weird because of settings and wavingSurfaceAmount
 		#endif
 		
 		
@@ -167,6 +180,14 @@ void main() {
 	materialId %= 100000;
 	
 	shadowcasterColor = getShadowcasterColor(ARG_IN);
+	
+	
+	#if WAVING_WATER_SURFACE_ENABLED == 1
+		#include "/import/at_tangent.glsl"
+		vec3 tangent = normalize(gl_NormalMatrix * at_tangent.xyz);
+		vec3 bitangent = normalize(cross(normal, tangent) * at_tangent.w);
+		tbn = mat3(tangent, bitangent, normal);
+	#endif
 	
 	
 	#if PHYSICALLY_WAVING_WATER_ENABLED == 1